Move to www.leagueoflegends.com domain
Patch notes are no longer local (like NA) and are on a www domain.

Along with this change we are no longer using a page-data json to load the patch notes, it is straight HTML on the patch notes page.

This changes it to using the page-data json that still contains data behind the patch notes.

With this though, the page-data.json still returns 200 for notes that don't exist. So we check the headers to see if it is returning json or text, if it is json it is a legit patch.